Removed debug messages

+++ chronos_server.h
@@ -237,7 +237,7 @@ typedef struct chronosServerContext_t
 
   
   /* The duration of one chronos experiment */
-  double duration_sec;
+  long duration_sec;
 
   int (*timeToDieFp)(void);
